+++ app/src/test/java/woowacourse/shopping/viewmodel/ProductDetailViewModelTest.kt
@@ -1,12 +1,13 @@
-package woowacourse.shopping
+package woowacourse.shopping.viewmodel
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
-import woowacourse.shopping.fixtures.InstantTaskExecutorExtension
-import woowacourse.shopping.fixtures.getOrAwaitValue
+import woowacourse.shopping.DummyShoppingRepository
 import woowacourse.shopping.productdetail.ProductDetailViewModel
+import woowacourse.shopping.viewmodel.fixtures.InstantTaskExecutorExtension
+import woowacourse.shopping.viewmodel.fixtures.getOrAwaitValue
 
 @ExtendWith(InstantTaskExecutorExtension::class)
 class ProductDetailViewModelTest {
