+++ app/src/test/java/woowacourse/shopping/viewmodel/ShoppingCartViewModelTest.kt
@@ -1,15 +1,16 @@
-package woowacourse.shopping
+package woowacourse.shopping.viewmodel
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
+import woowacourse.shopping.DummyShoppingRepository
 import woowacourse.shopping.domain.ShoppingCartItem
-import woowacourse.shopping.fixtures.InstantTaskExecutorExtension
-import woowacourse.shopping.fixtures.getOrAwaitValue
 import woowacourse.shopping.productlist.toProductUiModel
 import woowacourse.shopping.shoppingcart.ShoppingCartViewModel
+import woowacourse.shopping.viewmodel.fixtures.InstantTaskExecutorExtension
+import woowacourse.shopping.viewmodel.fixtures.getOrAwaitValue
 
 @ExtendWith(InstantTaskExecutorExtension::class)
 class ShoppingCartViewModelTest {
